multimaven: Move all code to maven package
